+++ pkg/monitoring/heartbeat/doc.go
@@ -1,0 +1,2 @@
+// Package heartbeat provides functionality for managing heartbeats in Opsgenie.
+package heartbeat